hrmm_backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
drf-spectacular>=0.28,<0.29
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.1,<2.0
PyYAML>=6.0,<7.0
